+++ src/services/dynamic-import.ts
@@ -6,12 +6,6 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-export const dynamicScript = async src => {
-  return new Promise((resolve, reject) => {
-    document.head.appendChild(Object.assign(document.createElement('script'), {
-      src,
-      onload: () => resolve(),
-      onerror: err => reject(err)
-    }));
-  });
-};
+
+// @ts-ignore
+export const dynamicImport = path => import(path);
